Fix AddCommand and AddCommandTest after reverted change and new commits.

+++ src/test/java/seedu/address/testutil/ModelStub.java
@@ -82,7 +82,7 @@ public class ModelStub implements Model {
     }
 
     @Override
-    public int totalModularCredits() {
+    public float totalModularCredits() {
         throw new AssertionError("This method should not be called.");
     }
